pyproject.toml
[project]
name = "med-xray-ssl"
version = "0.1.0"
description = "Fine-tuning pretrained SSL models on imbalanced chest X-rays for disease classification and detection"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "comet-ml==3.54.0",
    "einops==0.8.1",
    "hydra-core==1.3.2",
    "kagglehub==0.3.13",
    "matplotlib==3.10.7",
    "numpy==2.3.4",
    "pandas==2.3.3",
    "scikit-learn==1.7.2",
    "tensorboard==2.20.0",
    "torch==2.9.0",
    "torchmetrics==1.8.2",
    "torchvision==0.24.0",
    "tqdm==4.67.1",
    "transformers==4.57.1",
    "wandb==0.22.3",
]

[dependency-groups]
dev = [
    "pre-commit==4.3.0",
    "ruff==0.14.3",
    "ty==0.0.1a26",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["src"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff]
line-length = 100
